+++ cabal-testsuite/PackageTests/Exec/sandbox.test.hs
@@ -1,7 +1,7 @@
 import Test.Cabal.Prelude
 main = cabalTest $ do
     withSandbox $ do
-        fails $ cabal "exec" ["my-executable"]
-        cabal "install" []
-        cabal' "exec" ["my-executable"]
+        fails $ cabal "v1-exec" ["my-executable"]
+        cabal "v1-install" []
+        cabal' "v1-exec" ["my-executable"]
             >>= assertOutputContains "This is my-executable"